+++ CHANGELOG.md
@@ -5,6 +5,10 @@
 ### Changed
 
 - Requires `phpunit/php-code-coverage` `~14.2` to avoid using deprecated code
+- 
+### Fixed
+
+- `$assert->debug()` was leaking the last successful shrunk values
 
 ## 7.0.0 - 2026-05-14